requirements.txt
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.10
python-dotenv>=1.0
pandas>=2.2
numpy>=2.0
matplotlib>=3.9
pypdf>=5.0
beautifulsoup4>=4.12
markdown>=3.7
langchain>=1.0.0
langgraph>=1.0.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8
sentence-transformers>=3.0
httpx>=0.28
